+++ gamma_dashboard/dashboard/core/gamma/api/wrapper.py
@@ -80,14 +80,18 @@ class GammaApiWrapper:
 
         return self._send_request(absolute_url, **kwargs)
 
-    def get_leaderboard_info(self, username, user_signup_source, **kwargs):
+    def get_leaderboard_info(self, username, user_signup_source, course_id=None, **kwargs):
         """
         Return leaderboard data for user.
 
         Returns:
             dict: parsed leaderboard infromation.
         """
-        return self.request_api_endpoint('leaderboard', params={'username': username, 'signup_source': user_signup_source}, **kwargs)
+        return self.request_api_endpoint(
+            'leaderboard',
+            params={'username': username, 'signup_source': user_signup_source, 'course_id': course_id},
+            **kwargs
+        )
 
     def get_game_profile(self, username, **kwargs):
         """