+++ gamma_dashboard/static/dashboard/js/app/src/routes/AppRoutes.jsx
@@ -11,7 +11,7 @@ import { getMessages } from '../i18n/utils';
 import { getCookieByName } from '../utils';
 import { LeaderboardPage } from '../leaderboard';
 import { DashboardPage } from '../dashboard';
-import { URLS } from './constants';
+import { URLS, EXTERNAL_URLS } from './constants';
 
 const AppRoutes = () => {
   const locale = getCookieByName('openedx-language-preference') || 'en';
@@ -23,6 +23,7 @@ const AppRoutes = () => {
         <Routes>
           <Route exact path={URLS.dashboardPage} element={<DashboardPage />} />
           <Route exact path={URLS.leaderboardPage} element={<LeaderboardPage />} />
+          <Route path={EXTERNAL_URLS.leaderboardTabPage} element={<LeaderboardPage />} />
         </Routes>
       </Router>
     </IntlProvider>