+++ gamma_dashboard/static/dashboard/js/app/src/routes/constants.js
@@ -1,6 +1,10 @@
-const BASE_ROOT = '/gamma_dashboard';
+import { BASE_ROOT } from '../constants';
+
 export const URLS = {
   dashboardPage: `${BASE_ROOT}/dashboard`, // Dashboard page
   leaderboardPage: `${BASE_ROOT}/leaderboard`, // Leaderboard page
-  leaderboardTapPage: '/courses/:courseId/course_leaderboard', // Leaderboard tab in the course
+};
+
+export const EXTERNAL_URLS = {
+  leaderboardTabPage: '/courses/:courseId/course_leaderboard', // Course Leaderboard tab
 };